cc-oci-runtime: Autospec creation for version 2.1.0
The cc-proxy.socket symlink was installed to a bad location, but in
order to resolve the issue, it needs to be deleted first, then installed
to the correct location after the next format bump. This is due to SWUPD
not being able to handle directory -> symlink type changes at the
present time.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>